weather v0.0.1 Weather.TableFormatter

Functions

calculate_widths(dict)

Given a dictionary with key value pairs, calculate the maximum width needed for displaying keys and the maximum width needed for displaying values

Example

iex> dict = %{location: "Location 1", weather: "Weather 1", wind_pressure: "Wind Pressure 1", longitude: "Longitude 1", latitude: "Latitude 1"}
iex> Weather.TableFormatter.calculate_widths(dict)
[13, 15]

headers()

print_contents(dict)

Print contents, which are key value pairs, with padding

print_header(dict)

Given a key value pairs, print header with tile “Attributes” and “Values”, padded

print_separator(dict)

Print separator

print_table(dict)

Take a raw data in the form of a dictionary with keys and values. Print the header, titled “Attributes” and “Values”, print a separator, and print the data in a nicely format